Send "delete", not "reset", to session backend.
axiom_session exports delete/1, as does axiom_session_ets.
However, when axiom_session:delete is called, it sends
the "reset" message, which results in a call to the
(undefined) function axiom_session_ets:reset.
Changed the message from "reset" to "delete" to conform
to what appears to be a convention.