Fix relative redirects.
Cowboy stores the request host as a list of binaries, not as
a single binary or string, as Axiom has previously assumed.
This broke relative redirects when the host field was injected
as-is into the absolute redirect url ("." were omitted, so
if the host was "example.com" the redirect went to "examplecom"
due to iolist sematics).

Fixed relative redirects and added/fixed some test cases.